Making Spring AI Optional: - Separating AI Configuration from RedisModulesConfiguration into RedisAiConfiguration - Separating AI properties from RedisOMProperties into RedisOMAiProperties - Adding warning to BasicSentinelTest regarding running Redis manually being a requirement - Making dependencies related to AI features optional
