+++ demos/roms-vss/src/main/resources/application.properties
@@ -1,4 +1,4 @@
 spring.mvc.hiddenmethod.filter.enabled=true
 com.redis.om.vss.useLocalImages=false
 com.redis.om.vss.maxLines=300
-redis.om.spring.djl.enabled=true
+redis.om.spring.ai.djl.enabled=true
