+++ redis-om-spring/src/main/resources/META-INF/spring.factories
@@ -1,6 +1,7 @@
 # Auto Configure
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
 com.redis.om.spring.RedisModulesConfiguration
+com.redis.om.spring.RedisAiConfiguration
 org.springframework.boot.autoconfigure.AutoConfigurationImportFilter=\
 com.redis.om.spring.RedisRepositoriesExcludeFilter
 org.springframework.data.repository.core.support.RepositoryFactorySupport=com.redis.om.spring.repository.support.RedisEnhancedRepositoryFactory
