+++ redis-om-spring/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,2 +1,3 @@
 # Auto Configure
 com.redis.om.spring.RedisModulesConfiguration
+com.redis.om.spring.RedisAiConfiguration
