+++ redis-om-spring/src/test/resources/vss_on.yaml
@@ -1,5 +1,6 @@
 redis:
   om:
     spring:
-      djl:
-        enabled: true
+      ai:
+        djl:
+          enabled: true
